Move modules into components
