+++ Debate_RoomV2/Frontend/src/components/PeerTile.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useVideo } from '@100mslive/react-sdk';
-import Timer from '../modules/Timer';
+import Timer from './Timer';
 
 export default function PeerTile({
   peer,
