+++ Debate_RoomV2/Frontend/src/components/SpeakerQueue.jsx
